[CL-3361] Move proposals title to proposals page
