+++ front/app/containers/Admin/initiatives/index.tsx
@@ -4,22 +4,16 @@ import { Outlet as RouterOutlet } from 'react-router-dom';
 
 // components
 import HelmetIntl from 'components/HelmetIntl';
-import Button from 'components/UI/Button';
 import NavigationTabs, {
   Tab,
   TabsPageLayout,
 } from 'components/admin/NavigationTabs';
 import Link from 'utils/cl-router/Link';
-import { Box, Title } from '@citizenlab/cl2-component-library';
 
 // i18n
 import messages from './messages';
 import { WrappedComponentProps } from 'react-intl';
 import { injectIntl } from 'utils/cl-intl';
-
-// tracks
-import { trackEventByName } from 'utils/analytics';
-import tracks from './tracks';
 
 // styles
 import { InsertConfigurationOptions, ITab } from 'typings';
@@ -49,12 +43,6 @@ const InitiativesPage = memo<WrappedComponentProps & WithRouterProps>(
       },
     ]);
 
-    const onNewProposal = (pathname: string) => (_event) => {
-      trackEventByName(tracks.clickNewProposal.name, {
-        extra: { pathnameFrom: pathname },
-      });
-    };
-
     const handleData = (data: InsertConfigurationOptions<ITab>) =>
       setTabs(insertConfiguration<ITab>(data));
 
@@ -76,30 +64,6 @@ const InitiativesPage = memo<WrappedComponentProps & WithRouterProps>(
         </NavigationTabs>
 
         <StyledTabsPageLayout>
-          <Box
-            mb="30px"
-            width="100%"
-            display="flex"
-            justifyContent="space-between"
-            alignItems="center"
-            className="e2e-resource-header"
-          >
-            <Box>
-              <Title color="primary">
-                {formatMessage(messages.titleInitiatives)}
-              </Title>
-            </Box>
-            <Box ml="60px">
-              <Button
-                id="e2e-new-proposal"
-                buttonStyle="cl-blue"
-                icon="initiatives"
-                linkTo={`/initiatives/new`}
-                text={formatMessage(messages.addNewProposal)}
-                onClick={onNewProposal(pathname)}
-              />
-            </Box>
-          </Box>
           <HelmetIntl
             title={messages.metaTitle}
             description={messages.metaDescription}
