+++ front/app/containers/Admin/initiatives/manage/index.tsx
@@ -1,36 +1,69 @@
-import React, { PureComponent } from 'react';
-import { SectionTitle, SectionDescription } from 'components/admin/Section';
+import React from 'react';
+import { useLocation } from 'react-router-dom';
+import { SectionDescription } from 'components/admin/Section';
 
 import { FormattedMessage } from 'utils/cl-intl';
 import messages from '../messages';
-import styled from 'styled-components';
 import PostManager, { TFilterMenu } from 'components/admin/PostManager';
+import Button from 'components/UI/Button';
+import { Box, Title, Text } from '@citizenlab/cl2-component-library';
+import { trackEventByName } from 'utils/analytics';
+import tracks from './tracks';
+import { colors } from 'utils/styleUtils';
 
-const StyledDiv = styled.div`
-  margin-bottom: 30px;
-`;
+const InitiativesManagePage = () => {
+  const defaultFilterMenu = 'statuses';
+  const visibleFilterMenus: TFilterMenu[] = [defaultFilterMenu, 'topics'];
+  const { pathname } = useLocation();
 
-export default class InitiativesManagePage extends PureComponent {
-  render() {
-    const defaultFilterMenu = 'statuses';
-    const visibleFilterMenus: TFilterMenu[] = [defaultFilterMenu, 'topics'];
-    return (
-      <>
-        <StyledDiv>
-          <SectionTitle>
-            <FormattedMessage {...messages.titleManageTab} />
-          </SectionTitle>
-          <SectionDescription>
-            <FormattedMessage {...messages.subtitleManageTab} />
-          </SectionDescription>
-        </StyledDiv>
+  const onNewProposal = (pathname: string) => (_event) => {
+    trackEventByName(tracks.clickNewProposal.name, {
+      extra: { pathnameFrom: pathname },
+    });
+  };
+
+  return (
+    <Box width="100%">
+      <Box
+        mb="30px"
+        width="100%"
+        display="flex"
+        justifyContent="space-between"
+        alignItems="center"
+        className="e2e-resource-header"
+      >
+        <Box>
+          <Title color="primary">
+            <FormattedMessage {...messages.titleInitiatives} />
+          </Title>
+          <Text color="coolGrey600">
+            <FormattedMessage {...messages.subtitleDescription} />
+          </Text>
+        </Box>
+        <Box ml="60px">
+          <Button
+            id="e2e-new-proposal"
+            buttonStyle="cl-blue"
+            icon="initiatives"
+            linkTo={`/initiatives/new`}
+            text={<FormattedMessage {...messages.addNewProposal} />}
+            onClick={onNewProposal(pathname)}
+          />
+        </Box>
+      </Box>
+      <Box mb="30px" background={colors.white} p="40px">
+        <SectionDescription>
+          <FormattedMessage {...messages.subtitleDescription} />
+        </SectionDescription>
 
         <PostManager
           type="Initiatives"
           visibleFilterMenus={visibleFilterMenus}
           defaultFilterMenu={defaultFilterMenu}
         />
-      </>
-    );
-  }
-}
+      </Box>
+    </Box>
+  );
+};
+
+export default InitiativesManagePage;
