+++ front/app/containers/Admin/initiatives/messages.ts
@@ -25,14 +25,10 @@ export default defineMessages({
     id: 'app.containers.Admin.initiatives.settingsTabTitle',
     defaultMessage: 'Configure your proposals',
   },
-  titleManageTab: {
-    id: 'app.containers.Admin.initiatives.titleManageTab',
-    defaultMessage: 'Manage the initiatives on the platform.',
-  },
-  subtitleManageTab: {
-    id: 'app.containers.Admin.initiatives.subtitleManageTab',
+  subtitleDescription: {
+    id: 'app.containers.Admin.initiatives.subtitleDescription',
     defaultMessage:
-      'Add topics to initiatives, consult or change their statuses, officially update the citizens. Contact us if you want to use custom topics or statuses.',
+      'View the submitted proposals, change their status and add tags for easier filtering and categorization.',
   },
   warningTresholdSettings: {
     id: 'app.containers.Admin.initiatives.warningTresholdSettings',
